refactor(v1): move InitConfig into v1/configs
InitConfig lived in cli/init.py while every other entrypoint config (eval/serve/validate)
lives in v1/configs. Move it to configs/init.py to match, and complete the configs package
re-export (init + the previously-missing validate).
