+++ verifiers/v1/configs/__init__.py
@@ -1,6 +1,8 @@
 """Config schemas for the v1 entrypoints — the objects the CLIs parse."""
 
 from verifiers.v1.configs.eval import EvalConfig
+from verifiers.v1.configs.init import InitConfig
 from verifiers.v1.configs.serve import ServeConfig
+from verifiers.v1.configs.validate import ValidateConfig
 
-__all__ = ["EvalConfig", "ServeConfig"]
+__all__ = ["EvalConfig", "InitConfig", "ServeConfig", "ValidateConfig"]
